refactor: typing: Interaction Class seperation
* Create dedicated folder for all Interaction classes and subclasses
* Split interaction subclasses into dedicated files
* Change import locations and imported items to import from new interaction folder and dedicated interactions file (for interaction subclass)
* Change type annotations to dedicated subclasses (unfinished)
