+++ nextcord/ui/modal.py
@@ -23,7 +23,8 @@ __all__ = (
 
 
 if TYPE_CHECKING:
-    from ..interactions import ClientT, Interaction
+    from ..interactions.base import ClientT
+    from ..interactions.modal_submit_interaction import ModalSubmitInteraction
     from ..state import ConnectionState
     from ..types.components import ActionRow as ActionRowPayload
     from ..types.interactions import (
@@ -211,7 +212,7 @@ class Modal:
         self.children.clear()
         self.__weights.clear()
 
-    async def callback(self, interaction: Interaction):
+    async def callback(self, interaction: ModalSubmitInteraction):
         """|coro|
 
         The callback that is called when the user press the submit button.
@@ -234,7 +235,7 @@ class Modal:
         """
         pass
 
-    async def on_error(self, error: Exception, interaction: Interaction) -> None:
+    async def on_error(self, error: Exception, interaction: ModalSubmitInteraction) -> None:
         """|coro|
 
         A callback that is called when an item's callback or :meth:`interaction_check`
@@ -254,7 +255,7 @@ class Modal:
         print(f"Ignoring exception in modal {self}:", file=sys.stderr)
         traceback.print_exception(error.__class__, error, error.__traceback__, file=sys.stderr)
 
-    async def _scheduled_task(self, interaction: Interaction):
+    async def _scheduled_task(self, interaction: ModalSubmitInteraction):
         data: ModalSubmitInteractionData = interaction.data  # type: ignore
         for child in self.children:
             for component_data in _walk_component_interaction_data(data["components"]):
@@ -292,7 +293,7 @@ class Modal:
         asyncio.create_task(self.on_timeout(), name=f"discord-ui-modal-timeout-{self.id}")
         self.__stopped.set_result(True)
 
-    def _dispatch(self, interaction: Interaction):
+    def _dispatch(self, interaction: ModalSubmitInteraction):
         if self.__stopped.done():
             return
 
@@ -411,7 +412,7 @@ class ModalStore:
             if value is modal:
                 del self._modals[key]
 
-    def dispatch(self, custom_id: str, interaction: Interaction[ClientT]) -> None:
+    def dispatch(self, custom_id: str, interaction: ModalSubmitInteraction) -> None:
         self.__verify_integrity()
 
         key = (interaction.user.id, custom_id)  # type: ignore
