+++ nextcord/ui/view.py
@@ -304,7 +304,7 @@ class View:
         self.children.clear()
         self.__weights.clear()
 
-    async def interaction_check(self, interaction: Interaction) -> bool:
+    async def interaction_check(self, interaction: ViewInteraction) -> bool:
         """|coro|
 
         A callback that is called when an interaction happens within the view
@@ -339,7 +339,7 @@ class View:
         """
         pass
 
-    async def on_error(self, error: Exception, item: Item, interaction: Interaction) -> None:
+    async def on_error(self, error: Exception, item: Item, interaction: ViewInteraction) -> None:
         """|coro|
 
         A callback that is called when an item's callback or :meth:`interaction_check`
@@ -359,7 +359,7 @@ class View:
         print(f"Ignoring exception in view {self} for item {item}:", file=sys.stderr)
         traceback.print_exception(error.__class__, error, error.__traceback__, file=sys.stderr)
 
-    async def _scheduled_task(self, item: Item, interaction: Interaction):
+    async def _scheduled_task(self, item: Item, interaction: ViewInteraction):
         try:
             if self.timeout:
                 self.__timeout_expiry = time.monotonic() + self.timeout
@@ -395,7 +395,7 @@ class View:
         asyncio.create_task(self.on_timeout(), name=f"discord-ui-view-timeout-{self.id}")
         self.__stopped.set_result(True)
 
-    def _dispatch_item(self, item: Item, interaction: Interaction):
+    def _dispatch_item(self, item: Item, interaction: ViewInteraction):
         if self.__stopped.done():
             return
 
@@ -526,7 +526,7 @@ class ViewStore:
                 break
 
     def dispatch(
-        self, component_type: int, custom_id: str, interaction: Interaction[ClientT]
+        self, component_type: int, custom_id: str, interaction: ViewInteraction
     ) -> None:
         self.__verify_integrity()
         message_id: Optional[int] = interaction.message and interaction.message.id
